Improve dossiers and textes models
